edms-deployment-20251224-080728/backend/requirements/test.txt
# Minimal requirements for testing
Django==4.2.7
djangorestframework==3.14.0
django-extensions==3.2.3
celery==5.3.4
python-decouple==3.8